ch(backend): refactor my codes[Finishes #168902774]
