+++ Server/helpers/employeeValidation.js
@@ -5,12 +5,12 @@ const validation = (employee)=>{
         id:         Hapi.number().required(),
         firstname:  Hapi.string().required(),
         lastname:   Hapi.string().required(),
-        email:      Hapi.string().required(),
+        email:      Hapi.string().email().required(),
         password:   Hapi.string().required(),
         gender:     Hapi.string().required(),
         jobRole:    Hapi.string().required(),
         department: Hapi.string().required(),
-        address:    Hapi.string().required(),   
+        address:    Hapi.string().required(),  
     });
 
     return Schema.validate(employee);
