+++ Server/routes/auth.route.js
@@ -3,7 +3,7 @@ import userController from '../controllers/employeesController';
 
 const routes = Router();
 
-routes.post('/api/v1/auth/signup', userController.createAccount);
-routes.post('/api/v1/auth/signin', userController.login);
+routes.post('/auth/signup', userController.createAccount);
+routes.post('/auth/signin', userController.login);
 
 export default routes;
