+++ Server/routes/comments.route.js
@@ -4,7 +4,7 @@ import isLoggedIn from '../middlewares/isLoggedIn';
 
 const routes = Router();
 
-routes.post('/api/v1/articles/:articleId/comments', isLoggedIn, 
+routes.post('/articles/:articleId/comments', isLoggedIn, 
                             commentsController.createComment);
 
 export default routes;
