+++ package.json
@@ -2,10 +2,10 @@
   "name": "teamwork",
   "version": "1.0.0",
   "description": "Teamwork is an internal social network for organizations employees, the goal of this application is to facilitate more interaction between colleagues and facilitate team bonding.",
-  "main": "server.js",
+  "main": "app.js",
   "scripts": {
-    "start": "nodemon --exec babel-node ./server/server.js",
-    "dev": "nodemon --exec babel-node ./server/server.js",
+    "start": "nodemon --exec babel-node ./server/app.js",
+    "dev": "nodemon --exec babel-node ./server/app.js",
     "test": "nyc --require @babel/register mocha ./Server/test/*.js --exit",
     "coveralls": "nyc report --reporter=text-lcov | coveralls"
   },
